Integration tests for the Builder HTTP API
We had a suite of integration tests for the Builder API but it had some
problems. It destroyed the existing PG installation, it was slow to
start up and tear down, and it didn't cover many endpoints.

This aims to rectify these problems as much as possible. It uses pg_tmp
to create a temporary PG server just for the test run, which cleans up
after itself when it exits. It covers as many endpoints as possible. It
provides a way to run all of this stuff in an automated way, so we can
hopefully connect this up to Travis.

It's also worth noting that I converted the tests from TypeScript to
JavaScript since we weren't using any of the features of TypeScript and
it was just another layer of indirection to figure out when tests
failed.

Signed-off-by: Josh Black <raskchanky@gmail.com>